Valentines Day commit:
Havn't done much, but the background of the MusicItem cards is set, and
the display flex item accepts 4 items on one row.

Had the idea of putting the MusicItem cards on "the next page" of the website,
just like the about me page on my portfolio. With a nice image as background,
or maybe something with gradient...