Kind of backup commit:
Done:
Added artist data, made subpages which displays on 1 the cds, and on 2 the
artists. Also created the pageindicator component with one back and one
forward button. Now trying to make the buttons listen to a v-on:click event.

Just added a json file to move all the data from the Home.vue page to this, so
it can be fetched with an api.