Some progress made:
CDinfo, Artistinfo and LPinfo is displaying on all of the three pages in Home.vue,
Added a title above the musicItems section which dynamically gets a message based on the current page,
based on if statements.
Based on another set of if statements, each page has also a dynamically generated titleBubble, which is sticky.
Lp info added.
Background in navigation on product and artist page.