Monday:
Besides applying for a job... Also started with making the artist page with
parallax scrolling images. Did smth with routing that it takes the id of the
iten you clicked on the home page, so it can display the correct data. (Can,
hasn't happend yet - start that soon)

Good progress today :)