Since yesterday...
The product page gets the data from the clicked item displayed.
Used for this is: a created function with array.flat(2), array.filter and array.pop (pop!).
Remember! This doesn't work for the artists yet.
The design for the information (first section) on the product page is done.
Added some more high quality images so it looks better.

Made a list (on paper) with a couple [:)] of features to add to the site.

Nice work <3