Some awesome Spotify progress made:
Used Spotify api to get albums from their database, and display
it in the products page. Use in spotify app the URI to get the data
from that specific item.
Made more routes; artists, cds, lps, shoppingcart, and made it work
in the navigation.
Did some very basic styling to the shoppingcart page; for next time
if continueing working on the shoppingcart page, follow the course
on https://vueschool.io/lessons/meet-vuex.

:)