Tuesday the 3rd of march:
Achieved; a lot.
Contains:
Getting the data from spotify albums and displaying them in the a discover
cd's page - this page is with a new route, different than the product page with
the data from the json.
And adding a lot of Spotify URI's in the cds.json file.

Idea for next time; limit the amount of spotify items on the discover cd's page,
because it's already a lot and getiing more.

;)