Footer, SpotifyProductPage:
Created and finished the footer on the homepage.
Got the data from the clicked item on the cd's page to the spotifyproduct page,
and started designing it.
For next time; tracks. Create a for loop foreach item in the track, and create
foreach item in the tracklist a li item in a ul.

:)